# sila-me/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: SILA Cloud shared backend API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents/invoices:
    post:
      operationId: uploadInvoiceDocument
      tags: [invoices]
      summary: Upload an invoice document for server-side processing
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, minLength: 1, maxLength: 200 }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: "#/components/schemas/InvoiceUploadInput" }
      responses:
        "201":
          description: Invoice document captured
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvoiceUploadResponse" }
        "400":
          description: Invalid upload or missing idempotency key
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Scope, permission, or feature flag denied
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "413":
          description: File exceeds the configured upload limit
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "415":
          description: Unsupported invoice file type
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "502":
          description: SharePoint upload failure
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /invoices/{invoiceId}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Return an invoice and its extracted lines
      parameters:
        - { name: invoiceId, in: path, required: true, schema: { type: integer, minimum: 1 } }
      responses:
        "200":
          description: Invoice detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvoiceDetailResponse" }
        "404":
          description: Invoice not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    put:
      operationId: reviewInvoice
      tags: [invoices]
      summary: Save reviewed or corrected invoice extraction
      parameters:
        - { name: invoiceId, in: path, required: true, schema: { type: integer, minimum: 1 } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvoiceReviewInput" }
      responses:
        "200":
          description: Invoice review saved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvoiceStatusResponse" }
        "404":
          description: Invoice not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /invoices/{invoiceId}/confirm:
    post:
      operationId: confirmInvoice
      tags: [invoices]
      summary: Confirm reviewed invoice data
      parameters:
        - { name: invoiceId, in: path, required: true, schema: { type: integer, minimum: 1 } }
      responses:
        "200":
          description: Invoice confirmed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvoiceStatusResponse" }
        "404":
          description: Invoice not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
      required: [error]
...
    InvoiceUploadResponse:
      type: object
      properties:
        documentId: { type: integer }
        invoiceId: { type: ["integer", "null"] }
        uploadStatus: { type: string }
        processingStatus: { type: string }
        sharePointStored: { type: boolean }
        nextAction: { type: string }
      required: [documentId, invoiceId, uploadStatus, processingStatus, sharePointStored, nextAction]
    InvoiceUploadInput:
      type: object
      properties:
        file:
          type: string
          format: binary
        propertyId: { type: string }
        storeId: { type: string }
      required: [file]
...
    InvoiceDetailResponse:
      type: object
      properties:
        header: { type: object, additionalProperties: true }
        lines:
          type: array
          items: { type: object, additionalProperties: true }
        extraction:
          oneOf:
            - { type: object, additionalProperties: true }
            - { type: "null" }
      required: [header, lines, extraction]
    InvoiceReviewInput:
      type: object
      properties:
        supplierName: { type: string }
        invoiceNumber: { type: string }
        invoiceDate: { type: string, format: date }
        poNumber: { type: string }
        currency: { type: string }
        netAmount: { type: string }
        taxAmount: { type: string }
        grossAmount: { type: string }
        lines:
          type: array
          items:
            type: object
            additionalProperties: true
      required: [invoiceNumber]
    InvoiceStatusResponse:
      type: object
      properties:
        invoiceId: { type: integer }
        processingStatus: { type: string }
        validationStatus: { type: string }
        confirmedAt: { type: string, format: date-time }
      required: [invoiceId, processingStatus, validationStatus]
...