# sila-me/lib/api-zod/src/generated/types/invoiceReviewInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SILA Cloud shared backend API
 * OpenAPI spec version: 0.2.0
 */
import type { InvoiceReviewInputLinesItem } from './invoiceReviewInputLinesItem';

export interface InvoiceReviewInput {
  supplierName?: string;
  invoiceNumber: string;
  invoiceDate?: Date;
  poNumber?: string;
  currency?: string;
  netAmount?: string;
  taxAmount?: string;
  grossAmount?: string;
  lines?: InvoiceReviewInputLinesItem[];
}